commit:
types.ts: added properties to MinesweeperGameProps, to handle left and right click. And added some commented code for future use.

MinesweeperGame.tsx: added the onCellContext property, to handle the right click for flags, to interact with the backend.

Minesweeper.css: adjusted grid sizing and added also for different screen sizes.

MainPage.tsx: deleted unnecessary comments. added some commented code for future use. And added a way to interact with the backend to differentiate when there's a left or right click, setting the action to "reveal" for left click and "flag" for right click.

GameBoard.css: deleted width and height, because it was unformating the cells on reveal.

App.tsx: deleted unnecessary commented code.